requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.4
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
djangorestframework==3.15.1
easy-thumbnails==2.8.5
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-gettext==5.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
